+++ app/components/layout/main-navigation/profile-button/profile-button.component.ts
@@ -1,8 +1,8 @@
 import { ChangeDetectionStrategy, ChangeDetectorRef, Component, NgZone, OnDestroy, OnInit } from "@angular/core";
 import { Router } from "@angular/router";
 import { I18nService, Locale, LocaleService, TranslatedLocales } from "@batch-flask/core";
-import {  ElectronRemote, ElectronShell, FileSystemService } from "@batch-flask/ui";
 import { AutoUpdateService, UpdateStatus } from "@batch-flask/electron";
+import { ElectronRemote, ElectronShell, FileSystemService } from "@batch-flask/ui";
 import {
     ContextMenu, ContextMenuItem, ContextMenuSeparator, ContextMenuService, MultiContextMenuItem,
 } from "@batch-flask/ui/context-menu";
@@ -15,6 +15,7 @@ import { Constants } from "common";
 import * as path from "path";
 import { Subscription } from "rxjs";
 
+import { ProgressInfo } from "builder-util-runtime";
 import "./profile-button.scss";
 
 @Component({
@@ -30,6 +31,8 @@ export class ProfileButtonComponent implements OnDestroy, OnInit {
 
     private _currentUserSub: Subscription;
     private _updateSub: Subscription;
+    private _progressSub: Subscription;
+    private _downloadProgress: ProgressInfo;
 
     constructor(
         adalService: AdalService,
@@ -59,6 +62,19 @@ export class ProfileButtonComponent implements OnDestroy, OnInit {
             this.updateStatus = status;
             this.changeDetector.markForCheck();
         });
+
+        this._progressSub = this.autoUpdateService.downloadProgress.subscribe((downloadProgress) => {
+            this._downloadProgress = downloadProgress;
+            this.changeDetector.markForCheck();
+        });
+    }
+
+    public get downloadProgressTitle() {
+        if (this._downloadProgress) {
+            return `${this._downloadProgress.percent}%  ${this._downloadProgress.bytesPerSecond}bps`;
+        } else {
+            return null;
+        }
     }
 
     public ngOnInit() {
@@ -68,6 +84,7 @@ export class ProfileButtonComponent implements OnDestroy, OnInit {
     public ngOnDestroy() {
         this._currentUserSub.unsubscribe();
         this._updateSub.unsubscribe();
+        this._progressSub.unsubscribe();
     }
 
     public openSettingsContextMenu() {
