+++ src/@batch-flask/electron/electron-renderer.module.ts
@@ -1,7 +1,6 @@
 import { NgModule } from "@angular/core";
-import { AutoUpdateService, AutoUpdateRendererService } from "./auto-update";
-import { SharedServiceInjector } from "./shared-services-injector/shared-service-injector";
-
+import { AutoUpdateRendererService, AutoUpdateService } from "./auto-update";
+import { SharedServiceInjector } from "./shared-service-injector";
 
 /**
  * Module that contains electron service to be used in the renderer(Or browser)
